@@ -14,8 +14,9 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
+    <maven.compiler.source>16</maven.compiler.source>
+    <maven.compiler.target>16</maven.compiler.target>
+    <java.version>16</java.version>
     <sonar.projectKey>FerociousGaming_FirstTopDownGame</sonar.projectKey>
     <sonar.organization>ferociousgaming</sonar.organization>
     <sonar.host.url>https://sonarcloud.io</sonar.host.url>
