@@ -1,8 +1,9 @@
 package com.ferociousgaming.app;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+/*https://junit.org/junit5/docs/current/user-guide/index.html#writing-tests*/
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 /**
  * Unit test for simple App.
